Add BasicLocalSavegames (basic_feature) (#122)
* Add `basic_feature.BasicLocalSavegames`
(profile specific save games)

* Refactor Black & White 2 with BasicLocalSavegames

* Refactor Valheim with BasicLocalSavegames

* Refactor Vampire Bloodlines with BasicLocalSavegames

* Add info about basic local save games to readme
