+++ README.md
@@ -179,7 +179,8 @@ The meta-plugin provides some useful extra feature:
 2. **Basic save game preview:** If you use the Python version, and if you can easily obtain a picture (file)
   for any saves, you can provide basic save-game preview by using the `BasicGameSaveGameInfo`.
   See [games/game_witcher3.py](games/game_witcher3.py) for  more details.
-3. **Basic mod data checker** (Python):
+3. **Basic local save games** (Python): profile specific save games, as in [games/game_valheim.py](games/game_valheim.py).
+4. **Basic mod data checker** (Python):
   Check and fix different mod archive layouts for an automatic installation with the proper
   file structure, using simple (glob) patterns via `BasicModDataChecker`.
   See [games/game_valheim.py](games/game_valheim.py) and [game_subnautica.py](games/game_subnautica.py) for an example.
