+++ basic_features/__init__.py
@@ -1,4 +1,10 @@
+from .basic_local_savegames import BasicLocalSavegames
 from .basic_mod_data_checker import BasicModDataChecker, GlobPatterns
 from .basic_save_game_info import BasicGameSaveGameInfo
 
-__all__ = ["BasicModDataChecker", "BasicGameSaveGameInfo", "GlobPatterns"]
+__all__ = [
+    "BasicModDataChecker",
+    "BasicGameSaveGameInfo",
+    "GlobPatterns",
+    "BasicLocalSavegames",
+]
